requirements.txt
beautifulsoup4==4.13.4
boto3==1.36.13
python-dotenv==1.1.1
Requests==2.32.4
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
pydantic==2.11.7
pydantic_core==2.33.2